requirements.txt
langgraph==0.0.52
langchain-openai==0.1.0
langchain-google-genai==0.0.13
langchain-core==0.1.40
gradio==4.24.0
pillow==10.2.0
python-dotenv==1.0.0
openai==1.12.0
google-generativeai==0.3.2